Set Methods added to Storage class And also some Tests. four types of set Methods added 1. set(int[], Storage) 2. set(int[], float) 3. setExact(int[], float) 4. setFlat(int, float)
these methods added, the can be used to assign a value to the array.

1. set(int[] index, Storage val)
this methods helps to assign Storage values to another storage. if the assignable storage can be broadcast to the the underline shape of second Storage, it broadcast the value then set to it.
it works by iterating through items from offset to length, then calls it setFlat method at number 4.

2. set(int[] index, float val)
this is the same as the above but it assign float value instead of Storage. the shape can be any dim, inbound with the storage shape.

3. setExact(int[] index, float val)
this methods helps to set one value to the storage.
the index length must be equal to the shape length, also values must bounded to the shape of the storage.

4. setFlat(int index,float shape)
this method it the main set method used by other set methods.
the length ranges from (offset to offset + length).

some Tests available in Main.class
