--- src/gss/math/AbsStorage.java
@@ -1,8 +1,0 @@
-package gss.math;
-
-public interface AbsStorage
-{
-	public AbsStorage get(int...index);
-	public float getFloat(int...index);
-	public float getFlat(int index);
-}
